requirements.txt
beautifulsoup4
requests
selenium
webdriver-manager
python-dotenv